Fix CI: Int/Double mismatch in GroceryItemRow quantity edit dialog
item.grams is Int but EditGroceryQuantityDialog expects Double throughout
the edit chain (onEditQuantity, editManualQuantity, updateGrams).

+++ scan-eat-android/app/src/main/java/fr/scanneat/presentation/grocery/components/GroceryItemRow.kt
@@ -133,7 +133,7 @@ internal fun GroceryItemRow(
     if (showEditQuantity) {
         EditGroceryQuantityDialog(
             itemName = item.name,
-            initialGrams = item.grams,
+            initialGrams = item.grams.toDouble(),
             onConfirm = { grams -> onEditQuantity(grams); showEditQuantity = false },
             onDismiss = { showEditQuantity = false },
         )
